auctioneer: query unallocated tasks from ccu_store after receiving a START-TEST msg Moving task and tasklot to model module taskstatus is linked to tasklot
